get constraint to mod code
